+++ src/databricks/labs/ucx/source_code/languages.py
@@ -1,8 +1,8 @@
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.code.base import Fixer, Linter, SequentialLinter
-from databricks.labs.ucx.code.pyspark import SparkSql
-from databricks.labs.ucx.code.queries import FromTable
+from databricks.labs.ucx.source_code.base import Fixer, Linter, SequentialLinter
+from databricks.labs.ucx.source_code.pyspark import SparkSql
+from databricks.labs.ucx.source_code.queries import FromTable
 from databricks.labs.ucx.hive_metastore.table_migrate import Index
