+++ src/databricks/labs/ucx/source_code/lsp.py
@@ -12,14 +12,14 @@ from urllib.parse import parse_qsl
 from databricks.labs.blueprint.logger import install_logger
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.code.base import (
+from databricks.labs.ucx.source_code.base import (
     Advice,
     Advisory,
     Convention,
     Deprecation,
     Failure,
 )
-from databricks.labs.ucx.code.languages import Languages
+from databricks.labs.ucx.source_code.languages import Languages
 from databricks.labs.ucx.hive_metastore.table_migrate import Index, MigrationStatus
 
 logger = logging.getLogger(__name__)
