+++ tests/unit/source_code/test_languages.py
@@ -1,8 +1,8 @@
 import pytest
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.code.base import Fixer, Linter
-from databricks.labs.ucx.code.languages import Languages
+from databricks.labs.ucx.source_code.base import Fixer, Linter
+from databricks.labs.ucx.source_code.languages import Languages
 from databricks.labs.ucx.hive_metastore.table_migrate import Index
 
 index = Index([])
